requirements.txt
Appium-Python-Client
pytest
python-dotenv
langchain
langchain-openai
langchain-core




uiautomator2-mcp-server
langchain-mcp-adapters
langgraph


fastapi
uvicorn
pydantic
streamlit